update downloads to .astro
