+++ src/components/astro-downloads/astro-downloads.css
@@ -1,0 +1,3 @@
+.astro-downloads {
+	display: inherit;
+}
