+++ src/layouts/docs/docs-layout.astro
@@ -102,5 +102,4 @@ const githubSourceURL = `${globalData.designSystem.repository}/tree/${globalData
 <script>
 import './docs-layout.client'
 import 'project:components/color-swab-custom-element.ts'
-import 'project:utils/client/astro-download.ts'
 </script>
