--- src/pages/downloads.md
@@ -1,52 +1,0 @@
----
-title: Downloads
-description: Associated Astro Design System downloads and documentation.
-layout: project:layouts/docs/docs-layout.astro
----
-
-## Astro Design System
-
-- Astro Design Tokens ([Git Repository](https://github.com/RocketCommunicationsInc/astro-design-tokens))
-- Astro Web Components Source Code ([Git Repository](https://github.com/RocketCommunicationsInc/astro))
-- Astro UXDS Figma Page ([Figma](https://www.figma.com/@astrouxds))
-- Astro Icons ([SVG](https://github.com/RocketCommunicationsInc/astro/tree/main/packages/web-components/src/icons) | [Figma](https://www.figma.com/community/file/1022883566772542677))
-- Astro HTML Starter Kit ([Git Repository](https://github.com/RocketCommunicationsInc/astro-starter-kits/tree/main/html))
-- Astro Angular Starter Kit ([Git Repository](https://github.com/RocketCommunicationsInc/astro-starter-kits/tree/main/angular))
-- Astro React Starter Kit ([Git Repository](https://github.com/RocketCommunicationsInc/astro-starter-kits/tree/main/react))
-- Astro Vue Starter Kit ([Git Repository](https://github.com/RocketCommunicationsInc/astro-starter-kits/tree/main/vue))
-
-## Astro Web Components Download
-
-- Latest version ( <a id="latest-download" hidden download> Download </a> )
-
-**Need a previous version?**
-
-<div id="previousVersions" hidden>
-  <select id="allVersions">
-  </select>
-  <a id="downloadButton">Download</a>
-</div>
-
-## Themes
-
-- Astro AG Grid Theme ([Git Repository](https://github.com/RocketCommunicationsInc/ag-grid-theme))
-- Astro Tailwind Theme ([Git Repository](https://github.com/RocketCommunicationsInc/astro-tailwind))
-
-## GRM Service UX Design
-
-- GRM Design Specifications ([pdf](/pdf/grm-specifications.pdf))
-- GRM Wireframes ([pdf](/pdf/grm-wireframes.pdf))
-- GRM Dashboard App Source Code ([Git Repository](https://github.com/RocketCommunicationsInc/grm-dashboard-react))
-- GRM Equipment Manager App Source Code ([Git Repository](https://github.com/RocketCommunicationsInc/grm-equipment-react))
-- GRM Schedule App Source Code ([Git Repository](https://github.com/RocketCommunicationsInc/grm-schedule-react))
-
-## TT&C Service UX Design
-
-- TT&C Design Specifications ([pdf](/pdf/ttc-specifications.pdf))
-- TT&C Wireframes ([pdf](/pdf/ttc-wireframes.pdf))
-- TT&C Monitor App Source Code ([Git Repository](https://github.com/RocketCommunicationsInc/ttc-monitor-react))
-
-## FDS Service UX Design
-
-- [FD Design Specifications (pdf)](/pdf/fds-specifications.pdf)
-- [FD Wireframes (pdf)](/pdf/fds-wireframes.pdf)
